+++ packages/Adapters/ClassicDTK/test/map/tstVolumeSourceMap4.cpp
@@ -58,11 +58,11 @@ Teuchos::RCP<const Teuchos::Comm<Ordinal> > getDefaultComm()
 //---------------------------------------------------------------------------//
 // FieldEvaluator Implementation.
 class MyEvaluator : 
-    public DataTransferKit::FieldEvaluator<int,DataTransferKit::FieldContainer<double> >
+    public DataTransferKit::FieldEvaluator<unsigned long int,DataTransferKit::FieldContainer<double> >
 {
   public:
 
-    MyEvaluator( const Teuchos::ArrayRCP<int>& geom_gids, 
+    MyEvaluator( const Teuchos::ArrayRCP<unsigned long int>& geom_gids, 
 		 const Teuchos::RCP< const Teuchos::Comm<int> >& comm )
 	: d_geom_gids( geom_gids )
 	, d_comm( comm )
@@ -72,7 +72,7 @@ class MyEvaluator :
     { /* ... */ }
 
     DataTransferKit::FieldContainer<double> evaluate( 
-	const Teuchos::ArrayRCP<int>& gids,
+	const Teuchos::ArrayRCP<unsigned long int>& gids,
 	const Teuchos::ArrayRCP<double>& coords )
     {
 	Teuchos::ArrayRCP<double> evaluated_data( gids.size() );
@@ -96,7 +96,7 @@ class MyEvaluator :
 
   private:
 
-    Teuchos::ArrayRCP<int> d_geom_gids;
+    Teuchos::ArrayRCP<unsigned long int> d_geom_gids;
     Teuchos::RCP< const Teuchos::Comm<int> > d_comm;
 };
 
@@ -120,8 +120,8 @@ TEUCHOS_UNIT_TEST( VolumeSourceMap, one_to_many_parallel)
 
   // Setup source geometry. Only on proc zero
   const int geom_dim = 3;
-  Teuchos::RCP<GeometryManager<Box,int> > source_geometry_manager;
-  Teuchos::RCP<FieldEvaluator<int,FieldType> > source_evaluator;
+  Teuchos::RCP<GeometryManager<Box,unsigned long int> > source_geometry_manager;
+  Teuchos::RCP<FieldEvaluator<unsigned long int,FieldType> > source_evaluator;
 
   if (global_comm->getRank() == 0) {
     const int num_geom = 4;
@@ -131,13 +131,13 @@ TEUCHOS_UNIT_TEST( VolumeSourceMap, one_to_many_parallel)
     geometry[2] = Box(-1.0, -1.0, 2.0, 1.0, 1.0, 3.0);
     geometry[3] = Box(-1.0, -1.0, 3.0, 1.0, 1.0, 4.0);
 
-    Teuchos::ArrayRCP<int> geom_gids(num_geom);
+    Teuchos::ArrayRCP<unsigned long int> geom_gids(num_geom);
     for ( int i = 0; i < num_geom; ++i )
     {
       geom_gids[i] = i;
     }
     
-    source_geometry_manager = Teuchos::rcp(new GeometryManager<Box,int>(geometry,geom_gids,source_comm,geom_dim));
+    source_geometry_manager = Teuchos::rcp(new GeometryManager<Box,unsigned long int>(geometry,geom_gids,source_comm,geom_dim));
 
     source_evaluator = Teuchos::rcp(new MyEvaluator(geom_gids,source_comm));
   }
@@ -195,7 +195,7 @@ TEUCHOS_UNIT_TEST( VolumeSourceMap, one_to_many_parallel)
     Teuchos::rcp( new FieldManager<FieldType>( target_field, target_comm ) );
 
   // Setup and apply the volume source mapping.
-  VolumeSourceMap<Box,int,FieldType> volume_source_map( 
+  VolumeSourceMap<Box,unsigned long int,FieldType> volume_source_map( 
     global_comm, geom_dim, true, 1.0e-6 );
   volume_source_map.setup( source_geometry_manager, target_coord_manager );
   volume_source_map.apply( source_evaluator, target_space_manager );
